+++ src/App.tsx
@@ -2,7 +2,7 @@ import Home from "./pages/Home";
 
 function App() {
   return (
-    <main>
+    <main className="h-screen overflow-hidden bg-slate-200 font-serif">
       <Home />
     </main>
   );
